move and group component specific code away from workflows folder and into separate modular components folders
