+++ src/eo_api/components/data_registry/routes.py
