+++ src/eo_api/components/spatial_aggregation/schemas/spatial.py
@@ -1,0 +1,16 @@
+from pydantic import BaseModel
+
+
+class AggregationMethod(StrEnum):
+    """Supported numeric aggregation methods."""
+
+    MEAN = "mean"
+    SUM = "sum"
+    MIN = "min"
+    MAX = "max"
+
+
+class SpatialAggregationConfig(BaseModel):
+    """Spatial aggregation config."""
+
+    method: AggregationMethod = AggregationMethod.MEAN
