+++ src/eo_api/components/temporal_aggregation/schemas/temporal.py
@@ -1,0 +1,26 @@
+from pydantic import BaseModel
+
+
+class AggregationMethod(StrEnum):
+    """Supported numeric aggregation methods."""
+
+    MEAN = "mean"
+    SUM = "sum"
+    MIN = "min"
+    MAX = "max"
+
+
+class PeriodType(StrEnum):
+    """Supported temporal period types."""
+
+    HOURLY = "hourly"
+    DAILY = "daily"
+    MONTHLY = "monthly"
+    YEARLY = "yearly"
+
+
+class TemporalAggregationConfig(BaseModel):
+    """Temporal rollup config."""
+
+    target_period_type: PeriodType
+    method: AggregationMethod = AggregationMethod.SUM
