--- src/eo_api/data_manager/__init__.py
@@ -1,4 +1,0 @@
-from . import routes as routes
-from . import services as services
-
-__all__ = ["routes", "services"]
