--- src/eo_api/data_manager/services/__init__.py
@@ -1,5 +1,0 @@
-from . import constants as constants
-from . import downloader as downloader
-from . import utils as utils
-
-__all__ = ["constants", "downloader", "utils"]
