+++ src/eo_api/workflows/schemas.py
@@ -6,78 +6,6 @@ from typing import Any
 from pydantic import BaseModel, Field, model_validator
 
 
-class FeatureSourceType(StrEnum):
-    """Supported feature source backends."""
-
-    GEOJSON_FILE = "geojson_file"
-    DHIS2_LEVEL = "dhis2_level"
-    DHIS2_IDS = "dhis2_ids"
-
-
-class AggregationMethod(StrEnum):
-    """Supported numeric aggregation methods."""
-
-    MEAN = "mean"
-    SUM = "sum"
-    MIN = "min"
-    MAX = "max"
-
-
-class PeriodType(StrEnum):
-    """Supported temporal period types."""
-
-    HOURLY = "hourly"
-    DAILY = "daily"
-    MONTHLY = "monthly"
-    YEARLY = "yearly"
-
-
-class FeatureSourceConfig(BaseModel):
-    """How to fetch features for spatial aggregation."""
-
-    source_type: FeatureSourceType
-    geojson_path: str | None = None
-    dhis2_level: int | None = None
-    dhis2_ids: list[str] | None = None
-    dhis2_parent: str | None = None
-    feature_id_property: str = "id"
-
-    @model_validator(mode="after")
-    def validate_by_source(self) -> "FeatureSourceConfig":
-        """Enforce required fields per source backend."""
-        if self.source_type == FeatureSourceType.GEOJSON_FILE and not self.geojson_path:
-            raise ValueError("geojson_path is required when source_type='geojson_file'")
-        if self.source_type == FeatureSourceType.DHIS2_LEVEL and self.dhis2_level is None:
-            raise ValueError("dhis2_level is required when source_type='dhis2_level'")
-        if self.source_type == FeatureSourceType.DHIS2_IDS and not self.dhis2_ids:
-            raise ValueError("dhis2_ids is required when source_type='dhis2_ids'")
-        return self
-
-
-class TemporalAggregationConfig(BaseModel):
-    """Temporal rollup config."""
-
-    target_period_type: PeriodType
-    method: AggregationMethod = AggregationMethod.SUM
-
-
-class SpatialAggregationConfig(BaseModel):
-    """Spatial aggregation config."""
-
-    method: AggregationMethod = AggregationMethod.MEAN
-
-
-class Dhis2DataValueSetConfig(BaseModel):
-    """Mapping from aggregate outputs to DHIS2 DataValueSet fields."""
-
-    data_element_uid: str
-    category_option_combo_uid: str = "HllvX50cXC0"
-    attribute_option_combo_uid: str = "HllvX50cXC0"
-    data_set_uid: str | None = None
-    org_unit_property: str = "id"
-    stored_by: str | None = None
-
-
 class WorkflowExecuteRequest(BaseModel):
     """End-to-end workflow request."""
